Equality: Implement IEquatable interface

+++ tests/ValueWrapper.Tests.Integration.TestProject/TestNamespace1/RandomValue2.cs
@@ -2,7 +2,7 @@ using ValueWrapper.Attributes;
 
 namespace ValueWrapper.Tests.Integration.TestProject.TestNamespace1;
 
-[ValueWrapper(typeof(int))]
+[ValueWrapper(typeof(double))]
 public partial struct RandomValue2
 {
 }
